alpha-zero-general-chess-and-battlesnake-master/old_code/requirements.txt
tensorflow
tensorflow-gpu
keras
profilehooks
numpy
pyperclip
python-chess
h5py